lib: add parser callback to use the biodata-formats' COSMIC parser, #TASK-6427, #TASK-5318
On branch TASK-5318
Changes to be committed:
	modified:   cellbase-lib/src/main/java/org/opencb/cellbase/lib/builders/clinical/variant/CosmicIndexer.java
	new file:   cellbase-lib/src/main/java/org/opencb/cellbase/lib/builders/clinical/variant/CosmicIndexerCallback.java
